Add FailWorkflow convenience method to FSMContext
Move FSMContext into separate file instead of keeping in fsm_models.
